Move 'sdk' vs 'sdk-full' logic into convention plugins
