+++ gradle/plugins/build.gradle.kts
@@ -12,7 +12,7 @@ dependencies {
     implementation("net.ltgt.gradle:gradle-errorprone-plugin:3.1.0")
     implementation("org.gradlex:extra-java-module-info:1.8")
     implementation("org.gradlex:java-ecosystem-capabilities:1.5.1")
-    implementation("org.gradlex:java-module-dependencies:1.6.1")
+    implementation("org.gradlex:java-module-dependencies:1.6.2")
     implementation("org.sonarsource.scanner.gradle:sonarqube-gradle-plugin:4.4.1.3373")
 
     implementation("org.gradle.toolchains:foojay-resolver:0.7.0")
